requirements.txt
asgiref==3.4.0
Django==3.2.4
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
